requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
apsw==3.49.1.0
apswutils==0.0.2
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
click==8.1.8
deprecation==2.1.0
fastcore==1.7.29
fastlite==0.1.1
frozenlist==1.5.0
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
lxml==5.3.1
mistletoe==1.4.0
monsterui==1.0.10
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
postgrest==0.19.3
propcache==0.3.0
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.12.1
python-multipart==0.0.20
pyyaml==6.0.2
realtime==2.4.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
storage3==0.11.3
strenum==0.4.15
supabase==2.13.0
supafunc==0.9.3
typing-extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
yarl==1.18.3
